Create helper test file
