+++ tests/aksperiscopeintegration_test.go
@@ -76,14 +76,3 @@ func checkifpodsrunning(t *testing.T) bool {
 
 	return false
 }
-
-func TestFilterSliceElemsWithTest(t *testing.T) {
-	g := NewGomegaWithT(t)
-
-	crdNameContainsSmiTestPredicate := func(s string) bool { return strings.Contains(s, "smi-spec.io") }
-	testSlice := []string{"abc", "httproutegroups.specs.smi-spec.io", "def", "tcproutes.specs.smi-spec.io", "ghi"}
-	expectedSlice := []string{"httproutegroups.specs.smi-spec.io", "tcproutes.specs.smi-spec.io"}
-	actualSlice := utils.FilterSliceElemsWithTestPredicate(testSlice, crdNameContainsSmiTestPredicate)
-
-	g.Expect(actualSlice).To(Equal(expectedSlice))
-}
